chore: Add Heroicons Vue, TailwindCSS, Autoprefixer and PostCSS
This commit adds new dependencies to the package.json file, including Heroicons Vue, TailwindCSS, Autoprefixer, and PostCSS. It also includes the creation of two new files — postcss.config.js and tailwind.config.js with their corresponding code. In the main.ts file, the original imports were changed to include the correct path for the CSS file. Lastly, the entire CSS file was replaced with the contents from the TailwindCSS configuration.