test: add icon-pokedex.vue test
This commit adds a new test file to the src/components directory called "icon-pokedex.spec.ts". This file contains a test for the "Icon" component which checks if it renders the correct image. The test uses Vue test-utils and vitest library to mount the Icon component and assert that the "src" attribute of the "img" tag is "/src/assets/pokedex.svg".