feat: Added PokemonAPI and PokemonRepository classes
The commit adds two classes that interact with the PokeAPI to get a list of pokemon or their details. The `PokemonAPI` class uses Axios to get the list of pokemon or their details from the API. Meanwhile, the `PokemonRepository` class acts as a middleman that calls methods from `PokemonAPI` class to get a list of pokemon or a single pokemon.