feat: Updated README to reflect changes to project
The README was updated to reflect that the project is now a Pokedex developed with Vue 3, Vite, Vitest and Tailwind CSS. Installation, execution, and test commands were added along with a list of technologies used. The author of the project was also included. Lastly, a license section was added stating that the project is licensed under the MIT license and refers readers to the LICENSE file for more details.